+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `help("")` extension to options and arguments as an alternative to passing the help as an argument ([#207](https://github.com/ajalt/clikt/issues/207))
 - Added `valueSourceKey` parameter to `option`
 - Added `check{}` extensions to options and arguments as an alternative to `validate`
+- Added `prompt` and `confirm` functions to `CliktCommand` that call the `TermUi` equivalents with the current console.
 
 ### Fixed
 - Hidden options will no longer be suggested as possible typo corrections. ([#202](https://github.com/ajalt/clikt/issues/202))